Move mpfr_floordiv and mpfr_mod to a new mpfr_supplemental module.
